+++ components/utils/sleep.ts
@@ -1,0 +1,8 @@
+const sleep = (time: number = 2000) =>
+  new Promise((resolve) => {
+    setTimeout(() => {
+      resolve('');
+    }, time);
+  });
+
+export default sleep;
