+++ pages/logged-in.tsx
@@ -9,6 +9,7 @@ import Router from 'next/router';
 import { isValidJwt } from '../global/utils/egoTokenUtils';
 import PageLayout from '../components/PageLayout';
 import getInternalLink from '../global/utils/getInternalLink';
+import Loader from '../components/Loader';
 
 const fetchEgoToken = () => {
   const { NEXT_PUBLIC_EGO_API_ROOT, NEXT_PUBLIC_EGO_CLIENT_ID } = getConfig();
@@ -42,31 +43,6 @@ const fetchEgoToken = () => {
       localStorage.removeItem(EGO_JWT_KEY);
       Router.push(getInternalLink({ path: LOGIN_PATH }));
     });
-};
-
-// TODO: this is a placeholder Loader
-const Loader = () => {
-  return (
-    <div
-      css={(theme) => css`
-        border: 14px solid ${theme.colors.grey_3};
-        border-top: 14px solid ${theme.colors.secondary_dark};
-        border-radius: 50%;
-        width: 120px;
-        height: 120px;
-        animation: spin 2s linear infinite;
-
-        @keyframes spin {
-          0% {
-            transform: rotate(0deg);
-          }
-          100% {
-            transform: rotate(360deg);
-          }
-        }
-      `}
-    />
-  );
 };
 
 const LoginLoaderPage = createPage({
